timer component 기능별 분리 및 재사용 작업
