+++ app/src/main/java/com/earlybird/earlybirdcompose/presentation/screen/timer/component/TimerDoneContent.kt
@@ -1,0 +1,78 @@
+package com.earlybird.earlybirdcompose.presentation.screen.timer.component
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.earlybird.earlybirdcompose.ui.theme.EarlyBirdTheme
+
+//타이머가 종료되었을 때 나오는 종료 버튼
+@Composable
+fun TimerDoneContent(
+    content: String,
+    buttonContent: String,
+    onClick: () -> Unit
+) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+        modifier = Modifier
+            .fillMaxSize()
+            .padding(top = 66.dp)
+    ){
+        Text(
+            text = "COMPLETE!",
+            color = EarlyBirdTheme.colors.mainBlue,
+            fontSize = 36.sp,
+            fontWeight = FontWeight.Bold
+        )
+        Spacer(modifier = Modifier.height(15.dp))
+
+        Text(
+            text = "우와! 우리가 해냈다\n다음에도 같이 하자!!",
+            color = EarlyBirdTheme.colors.fontBlack,
+            fontSize = 24.sp,
+            fontWeight = FontWeight.Bold
+        )
+        Spacer(modifier = Modifier.height(40.dp))
+        Button(
+            onClick = onClick,
+            modifier = Modifier
+                .width(218.dp)
+                .height(48.dp)
+                .shadow(
+                    elevation = 8.dp,
+                    shape = RoundedCornerShape(28.dp),
+                    spotColor = Color.Black.copy(alpha = 0.25f)
+                ),
+            shape = RoundedCornerShape(28.dp),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = EarlyBirdTheme.colors.mainBlue,
+                contentColor = EarlyBirdTheme.colors.white
+            )
+        ) {
+            Text(
+                text = "완료",
+                fontSize = 28.sp,
+                fontWeight = FontWeight.SemiBold,
+                textAlign = TextAlign.Center
+            )
+        }
+    }
+}
